// frontend/src/types/supabase.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      jobs: {
        Row: {
          applicant_count: number | null;
          description: string | null;
          employer_id: string;
          expires_at: string | null;
          job_id: string;
          job_type: string | null;
          location: string | null;
          max_experience: number | null;
          max_salary: number | null;
          min_experience: number | null;
          min_salary: number | null;
          posted_at: string | null;
          status: Database["public"]["Enums"]["job_status_enum"] | null;
          title: string;
        };
        Insert: {
          applicant_count?: number | null;
          description?: string | null;
          employer_id: string;
          expires_at?: string | null;
          job_id?: string;
          job_type?: string | null;
          location?: string | null;
          max_experience?: number | null;
          max_salary?: number | null;
          min_experience?: number | null;
          min_salary?: number | null;
          posted_at?: string | null;
          status?: Database["public"]["Enums"]["job_status_enum"] | null;
          title: string;
        };
        Update: {
          applicant_count?: number | null;
          description?: string | null;
          employer_id?: string;
          expires_at?: string | null;
          job_id?: string;
          job_type?: string | null;
          location?: string | null;
          max_experience?: number | null;
          max_salary?: number | null;
          min_experience?: number | null;
          min_salary?: number | null;
          posted_at?: string | null;
          status?: Database["public"]["Enums"]["job_status_enum"] | null;
          title?: string;
        };
        Relationships: [
          {
            foreignKeyName: "jobs_employer_id_fkey";
            columns: ["employer_id"];
            isOneToOne: false;
            referencedRelation: "employers";
            referencedColumns: ["user_id"];
          },
        ];
      };
// ...
    };
// ...
  };
};

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">;

type DefaultSchema = DatabaseWithoutInternals[Extract<
  keyof Database,
  "public"
>];
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals;
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals;
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never;